Fixes some issues and replaces most of stock content for for actual content
