+++ components/ui/Divider.js
@@ -1,7 +1,8 @@
 import styled from "styled-components";
 
 const Divider = styled.hr`
-  border: 1px solid ${({ theme, color = "grey" }) => theme[color]};
+  border: ${({ thickness = "1px" }) => thickness} solid
+    ${({ theme, color = "darkGrey" }) => theme[color]};
   width: ${({ width = "100%" }) => width};
   margin: ${({ margin = "10px 0" }) => margin};
 `;
