+++ data/portfolio.json
@@ -1,28 +1,56 @@
 {
   "carpentry": [
     {
-      "title": "Hotel Eva",
+      "title": "Aeroporto Faro",
       "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
-      "images": ["/static/eva-faro.jpeg", "/static/hotel-tivoli.jpg"]
-    },
-    {
-      "title": "Hotel Tivoli",
-      "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
-      "images": ["/static/hotel-tivoli.jpg"]
+      "images": [
+        "/static/carpentry/aeroporto-faro/aeroporto-faro-2.jpg",
+        "/static/carpentry/aeroporto-faro/aeroporto-faro-1.jpg",
+        "/static/carpentry/aeroporto-faro/aeroporto-faro-3.jpg",
+        "/static/carpentry/aeroporto-faro/aeroporto-faro-4.jpg",
+        "/static/carpentry/aeroporto-faro/aeroporto-faro-5.jpg",
+        "/static/carpentry/aeroporto-faro/aeroporto-faro-6.jpg"
+      ]
     }
   ],
   "painting": [
     {
-      "title": "Hotel Eva",
+      "title": "Moradias Casa do Rio",
       "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
-      "images": ["/static/eva-faro.jpeg"]
+      "images": [
+        "/static/painting/moradias-casas-do-rio-alhandra/alhandra-1.jpg",
+        "/static/painting/moradias-casas-do-rio-alhandra/alhandra-2.jpg",
+        "/static/painting/moradias-casas-do-rio-alhandra/alhandra-4.jpg",
+        "/static/painting/moradias-casas-do-rio-alhandra/alhandra-5.jpg",
+        "/static/painting/moradias-casas-do-rio-alhandra/alhandra-6.jpg",
+        "/static/painting/moradias-casas-do-rio-alhandra/alhandra-7.jpg",
+        "/static/painting/moradias-casas-do-rio-alhandra/alhandra-8.jpg"
+      ]
     }
   ],
   "remodeling": [
     {
-      "title": "Hotel Eva",
-      "description": "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.",
-      "images": ["/static/eva-faro.jpeg"]
+      "title": "Louroé Quartos",
+      "description": "",
+      "images": [
+        "/static/remodeling/louroe-quartos/louroe-quartos-14.jpg",
+        "/static/remodeling/louroe-quartos/louroe-quartos-1.jpg",
+        "/static/remodeling/louroe-quartos/louroe-quartos-2.jpg",
+        "/static/remodeling/louroe-quartos/louroe-quartos-3.jpg",
+        "/static/remodeling/louroe-quartos/louroe-quartos-4.jpg",
+        "/static/remodeling/louroe-quartos/louroe-quartos-5.jpg",
+        "/static/remodeling/louroe-quartos/louroe-quartos-6.jpg",
+        "/static/remodeling/louroe-quartos/louroe-quartos-7.jpg",
+        "/static/remodeling/louroe-quartos/louroe-quartos-8.jpg",
+        "/static/remodeling/louroe-quartos/louroe-quartos-9.jpg",
+        "/static/remodeling/louroe-quartos/louroe-quartos-10.jpg",
+        "/static/remodeling/louroe-quartos/louroe-quartos-11.jpg",
+        "/static/remodeling/louroe-quartos/louroe-quartos-12.jpg",
+        "/static/remodeling/louroe-quartos/louroe-quartos-13.jpg",
+        "/static/remodeling/louroe-quartos/louroe-quartos-15.jpg",
+        "/static/remodeling/louroe-quartos/louroe-quartos-16.jpg",
+        "/static/remodeling/louroe-quartos/louroe-quartos-17.jpg"
+      ]
     }
   ]
 }
